Extract and modularise `ui/reader` to `reader-ui`
This removes Reader state from `ui`, while avoiding all of Reader being
loaded on boot.

Also extracts a few inline selectors into a proper selector module.
