+++ client/blocks/reader-post-card/index.jsx
@@ -30,7 +30,7 @@ import {
 	getSourceFollowUrl as getDiscoverFollowUrl,
 } from 'reader/discover/helper';
 import DiscoverFollowButton from 'reader/discover/follow-button';
-import { expandCard as expandCardAction } from 'state/ui/reader/card-expansions/actions';
+import { expandCard as expandCardAction } from 'state/reader-ui/card-expansions/actions';
 import isReaderCardExpanded from 'state/selectors/is-reader-card-expanded';
 
 /**
