+++ client/reader/sidebar/reader-sidebar-organizations/list.jsx
@@ -13,8 +13,8 @@ import { connect } from 'react-redux';
 import ExpandableSidebarMenu from 'layout/sidebar/expandable';
 import ReaderSidebarOrganizationsListItem from './list-item';
 import getOrganizationSites from 'state/reader/follows/selectors/get-reader-follows-organization';
-import { toggleReaderSidebarOrganization } from 'state/ui/reader/sidebar/actions';
-import { isOrganizationOpen } from 'state/ui/reader/sidebar/selectors';
+import { toggleReaderSidebarOrganization } from 'state/reader-ui/sidebar/actions';
+import { isOrganizationOpen } from 'state/reader-ui/sidebar/selectors';
 import { AUTOMATTIC_ORG_ID } from 'state/reader/organizations/constants';
 import ReaderSidebarHelper from 'reader/sidebar/helper';
 import SidebarItem from 'layout/sidebar/item';
