+++ client/state/reader-ui/card-expansions/reducer.js
@@ -1,7 +1,6 @@
 /**
  * Internal dependencies
  */
-
 import { withoutPersistence } from 'state/utils';
 import { READER_EXPAND_CARD, READER_RESET_CARD_EXPANSIONS } from 'state/reader/action-types';
 import { keyToString } from 'reader/post-key';
