+++ client/state/reader-ui/seen-posts/selectors.js
@@ -1,9 +1,14 @@
 /**
+ * Internal dependencies
+ */
+import 'state/reader-ui/init';
+
+/**
  * Check if we have Reader unseen content
  *
  * @param state redux state
  * @returns {boolean} whether or not we have unseen posts
  */
 export const hasUnseen = ( state ) => {
-	return state.ui.reader.hasUnseenPosts;
+	return state.readerUi.hasUnseenPosts;
 };
