+++ client/state/reader-ui/sidebar/reducer.js
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import { xor } from 'lodash';
+
 /**
  * Internal dependencies
  */
