+++ client/state/selectors/is-reader-card-expanded.js
@@ -1,10 +1,11 @@
 /**
  * Internal dependencies
  */
-
 import { keyToString } from 'reader/post-key';
+
+import 'state/reader-ui/init';
 
 export default function isReaderCardExpanded( state, postKey ) {
 	const key = keyToString( postKey );
-	return !! ( key && state.ui.reader.cardExpansions[ key ] );
+	return !! ( key && state.readerUi.cardExpansions[ key ] );
 }
